pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "distributed-training-pipeline"
version = "1.0.0"
description = "Production-style Distributed Continuous Training Pipeline with Airflow, HuggingFace Accelerate, and MLflow"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "MLOps Team"}
]
keywords = ["mlops", "distributed-training", "airflow", "pytorch", "mlflow"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.1.0",
    "accelerate>=0.25.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.4.0",
    "mlflow>=2.10.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "PyYAML>=6.0.0",
    "structlog>=24.1.0",
    "tqdm>=4.66.0",
    "click>=8.1.0",
    "rich>=13.7.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
airflow = [
    "apache-airflow>=2.8.0",
    "apache-airflow-providers-celery>=3.5.0",
]
serving = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[project.scripts]
train = "src.training.distributed_train:main"
serve = "src.deployment.serving:main"

[project.urls]
Repository = "https://github.com/example/distributed-training-pipeline"
Documentation = "https://github.com/example/distributed-training-pipeline#readme"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ["py310", "py311"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src --cov-report=term-missing"
